python-backend/requirements.txt
opencv-python
numpy
flask
flask-cors
pillow